+++ pkgs/broken.nix
@@ -1,5 +1,4 @@
 [
-  "fcitx5-mozc-ext-neologd"
   "vrchat-vpm-cli"
   "fence-agents"
   "drbd9-dkms"