+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.3.1](https://github.com/misumisumi/flakes/compare/v1.3.0...v1.3.1) (2024-07-01)
+
+
+### Bug Fixes
+
+* remove vrc-get-latest ([407529c](https://github.com/misumisumi/flakes/commit/407529c3cbc871ecd8572264a3dbd492f618684f))
+
+## [1.3.0](https://github.com/misumisumi/flakes/compare/v1.2.2...v1.3.0) (2024-06-14)
+
+
+### Features
+
+* add julius-speech ([6c66300](https://github.com/misumisumi/flakes/commit/6c66300388dda17f5b2d5b2c078d3df373280382))
+* **moralerspace:** add moralerspace-font ([b417cfd](https://github.com/misumisumi/flakes/commit/b417cfd3deeb600e816baff4753830f755343209))
+
+
+### Bug Fixes
+
+* **fcitx5-mozc-ext-neologd:** remove mozc-ext-neologd ([6ca8dec](https://github.com/misumisumi/flakes/commit/6ca8decc99578844739150428039683f16a6479f))
+* **SpeechRecognition:** remove SpeechRecognition ([449f5a8](https://github.com/misumisumi/flakes/commit/449f5a83a931b02353a076f77015396f05cd4893))
+
+## [1.2.2](https://github.com/misumisumi/flakes/compare/v1.2.1...v1.2.2) (2024-06-02)
+
+
+### Bug Fixes
+
+* remove textlintMyEnv ([7f6bc5f](https://github.com/misumisumi/flakes/commit/7f6bc5fd3f1313951e9f23daa0bce8333fd8d601))
+
 ## [1.2.1](https://github.com/misumisumi/flakes/compare/v1.2.0...v1.2.1) (2024-04-12)